Make sure user endpoints also require authentication.
