+++ app/routers/authentication.py
@@ -1,5 +1,6 @@
 from bson import ObjectId
 from fastapi import APIRouter, Depends, HTTPException
+from passlib.hash import bcrypt
 from pymongo import MongoClient
 
 from app import dependencies
@@ -9,6 +10,15 @@ from app.auth import AuthHandler
 auth_handler = AuthHandler()
 
 router = APIRouter()
+
+
+@router.post("/users", response_model=UserDB)
+async def save_user(userIn: UserIn, db: MongoClient = Depends(dependencies.get_db)):
+    hashed_password = bcrypt.hash(userIn.password)
+    userDB = UserDB(**userIn.dict(), hashed_password=hashed_password)
+    res = await db["users"].insert_one(userDB.to_mongo())
+    found = await db["users"].find_one({"_id": res.inserted_id})
+    return UserDB.from_mongo(found).dict(exclude={"create_at"})
 
 
 @router.post("/login")
