+++ app/routers/users.py
@@ -1,23 +1,13 @@
 from typing import List
 
 from bson import ObjectId
-from fastapi import APIRouter, Request, HTTPException, Depends
+from fastapi import APIRouter, HTTPException, Depends
 from pymongo import MongoClient
 
 from app import dependencies
-from app.models.users import UserDB, UserIn
-from passlib.hash import bcrypt
+from app.models.users import UserDB
 
 router = APIRouter()
-
-
-@router.post("", response_model=UserDB)
-async def save_user(userIn: UserIn, db: MongoClient = Depends(dependencies.get_db)):
-    hashed_password = bcrypt.hash(userIn.password)
-    userDB = UserDB(**userIn.dict(), hashed_password=hashed_password)
-    res = await db["users"].insert_one(userDB.to_mongo())
-    found = await db["users"].find_one({"_id": res.inserted_id})
-    return UserDB.from_mongo(found).dict(exclude={"create_at"})
 
 
 @router.get("", response_model=List[UserDB])
